Diagrams.Backend.PGF.CmdLine

Description

Convenient creation of command-line-driven executables for rendering diagrams using the PGF backend.

General form of main

The mainWith method unifies all of the other forms of main and is now the recommended way to build a command-line diagrams program. It works as a direct replacement for defaultMain or multiMain as well as allowing more general arguments. For example, given a function that produces a diagram when given an Int and a Colour Double, mainWith will produce a program that looks for additional number and color arguments.

... definitions ...
f :: Int -> Colour Double -> Diagram PGF
f i c = ...

main = mainWith f

We can run this program as follows:

$ ghc --make mydiagram

# output image.tex built by `f 20 red`
$ ./MyDiagram -o image.tex -w 200 20 red

mainWith :: (Mainable d, Parseable (MainOpts d)) => d -> IO ()

Main entry point for command-line diagram creation. This is the method that users will call from their program main. For instance an expected user program would take the following form.

import Diagrams.Prelude
import Diagrams.Backend.TheBestBackend.CmdLine

d :: Diagram B R2
d = ...

main = mainWith d

Most backends should be able to use the default implementation. A different implementation should be used to handle more complex interactions with the user.

Supported forms of main

defaultMain :: Diagram PGF -> IO ()

mainWithSurf :: Surface -> Diagram PGF -> IO ()

Allows you to pick a surface the diagram will be rendered with. (This

onlineMain :: OnlineTex (Diagram PGF) -> IO ()

Same as defaultMain but takes an online pgf diagram.

onlineMainWithSurf :: Surface -> OnlineTex (Diagram PGF) -> IO ()

Same as mainWithSurf but takes an online pgf diagram.

multiMain :: [(String, Diagram PGF)] -> IO ()

module Diagrams.Backend.PGF

Orphan instances

Parseable TexFormat

Methods

parser :: Parser TexFormat

ToResult d => ToResult (OnlineTex d)

Associated Types

type Args (OnlineTex d) :: *

type ResultOf (OnlineTex d) :: *

Methods

toResult :: OnlineTex d -> Args (OnlineTex d) -> ResultOf (OnlineTex d)

TypeableFloat n => Mainable [(String, QDiagram PGF V2 n Any)]

Associated Types

type MainOpts [(String, QDiagram PGF V2 n Any)] :: *

Methods

mainArgs :: [(String, QDiagram PGF V2 n Any)] -> IO (MainOpts [(String, QDiagram PGF V2 n Any)])

mainRender :: MainOpts [(String, QDiagram PGF V2 n Any)] -> [(String, QDiagram PGF V2 n Any)] -> IO ()

mainWith :: [(String, QDiagram PGF V2 n Any)] -> IO ()

TypeableFloat n => Mainable (OnlineTex (QDiagram PGF V2 n Any))

Associated Types

type MainOpts (OnlineTex (QDiagram PGF V2 n Any)) :: *

Methods

mainArgs :: OnlineTex (QDiagram PGF V2 n Any) -> IO (MainOpts (OnlineTex (QDiagram PGF V2 n Any)))

mainRender :: MainOpts (OnlineTex (QDiagram PGF V2 n Any)) -> OnlineTex (QDiagram PGF V2 n Any) -> IO ()

mainWith :: OnlineTex (QDiagram PGF V2 n Any) -> IO ()

TypeableFloat n => Mainable (Surface, QDiagram PGF V2 n Any)

Associated Types

type MainOpts (Surface, QDiagram PGF V2 n Any) :: *

Methods

mainArgs :: (Surface, QDiagram PGF V2 n Any) -> IO (MainOpts (Surface, QDiagram PGF V2 n Any))

mainRender :: MainOpts (Surface, QDiagram PGF V2 n Any) -> (Surface, QDiagram PGF V2 n Any) -> IO ()

mainWith :: (Surface, QDiagram PGF V2 n Any) -> IO ()

TypeableFloat n => Mainable (Surface, OnlineTex (QDiagram PGF V2 n Any))

Associated Types

type MainOpts (Surface, OnlineTex (QDiagram PGF V2 n Any)) :: *

Methods

mainArgs :: (Surface, OnlineTex (QDiagram PGF V2 n Any)) -> IO (MainOpts (Surface, OnlineTex (QDiagram PGF V2 n Any)))

mainRender :: MainOpts (Surface, OnlineTex (QDiagram PGF V2 n Any)) -> (Surface, OnlineTex (QDiagram PGF V2 n Any)) -> IO ()

mainWith :: (Surface, OnlineTex (QDiagram PGF V2 n Any)) -> IO ()

TypeableFloat n => Mainable (QDiagram PGF V2 n Any)

Associated Types

type MainOpts (QDiagram PGF V2 n Any) :: *

Methods

mainArgs :: QDiagram PGF V2 n Any -> IO (MainOpts (QDiagram PGF V2 n Any))

mainRender :: MainOpts (QDiagram PGF V2 n Any) -> QDiagram PGF V2 n Any -> IO ()

mainWith :: QDiagram PGF V2 n Any -> IO ()