Diagrams.Backend.PGF.Render

Description

This is an internal module exposeing internals for rendering a diagram. This is for advanced use only. PGF has enought for general use.

Documentation

data PGF

This data declaration is simply used as a token to distinguish this rendering engine.

Constructors

PGF

data family Options b (v :: * -> *) n :: *

Backend-specific rendering options.

data family Render b (v :: * -> *) n :: *

An intermediate representation used for rendering primitives. (Typically, this will be some sort of monad, but it need not be.) The Renderable class guarantees that a backend will be able to convert primitives into this type; how these rendered primitives are combined into an ultimate Result is completely up to the backend.

Lenses

surface :: Lens' (Options PGF V2 n) Surface

Lens onto the surface used to render.

sizeSpec :: Lens' (Options PGF V2 n) (SizeSpec V2 n)

Lens onto the SizeSpec2D.

readable :: Lens' (Options PGF V2 n) Bool

Lens onto whether the lines of the TeX output are indented.

standalone :: Lens' (Options PGF V2 n) Bool

Lens onto whether a standalone TeX document should be produced.

Utilities

escapeString :: String -> String

Escapes some common characters in a string. Note that this does not mean the string can't create an error, it mearly escapes common characters.