Graphics.Rendering.PGF

Description

Interface to PGF. See the manual http://www.ctan.org/pkg/pgf for details.

Documentation

renderWith :: (RealFloat n, Typeable n) => Surface -> Bool -> Bool -> V2 n -> Render n -> Builder

type RenderM n m = RWS RenderInfo Builder (RenderState n) m

Type for render monad.

type Render n = RenderM n ()

Convenient type for building.

initialState :: (Typeable n, Floating n) => RenderState n

Starting state for running the builder.

Environments

scope :: Render n -> Render n

Wrap the rendering in a scope.

epsilon :: Fractional n => n

ε = 0.0001 is the limit at which lines are no longer stroked.

Lenses

style :: forall n. Lens' (RenderState n) (Style V2 n)

units

bp :: RealFloat a => a -> Render n

Render n length with bp (big point = 1 px at 72 dpi) units.

pt :: RealFloat a => a -> Render n

Render n length with pt units.

mm :: RealFloat a => a -> Render n

Render n length with mm units.

px :: RealFloat a => a -> Render n

Render n length with px units.

RenderM commands

ln :: Render n -> Render n

raw :: Builder -> Render n

rawString :: String -> Render n

pgf :: Builder -> Render n

bracers :: Render n -> Render n

Wrap a `Render n` in { .. }.

brackets :: Render n -> Render n

Wrap a `Render n` in [ .. ].

Paths

path :: RealFloat n => Path V2 n -> Render n

trail :: RealFloat n => Trail V2 n -> Render n

segment :: RealFloat n => Segment Closed V2 n -> Render n

usePath :: Bool -> Bool -> Render n

usePath fill stroke combined in one function.

lineTo :: RealFloat n => V2 n -> Render n

Move path by vector.

curveTo :: RealFloat n => V2 n -> V2 n -> V2 n -> Render n

Make curved path from vectors.

moveTo :: RealFloat n => P2 n -> Render n

Move path to point.

closePath :: Render n

Close the current path.

clip :: Render n

Use the defined path a clip for everything that follows in the current scope. Stacks.

stroke :: Render n

Stroke the defined path using parameters from current scope.

fill :: Render n

Fill the defined path using parameters from current scope.

asBoundingBox :: Render n

Uses the current path as the bounding box for whole picture.

Strokeing Options

setDash :: RealFloat n => Dashing n -> Render n

Sets the dash for the current scope. Must be done before stroking.

setLineWidth :: RealFloat n => n -> Render n

Sets the line width in current scope. Must be done before stroking.

setLineCap :: LineCap -> Render n

Sets the line cap in current scope. Must be done before stroking.

setLineJoin :: LineJoin -> Render n

Sets the line join in current scope. Must be done before stroking.

setMiterLimit :: RealFloat n => n -> Render n

Sets the miter limit in the current scope. Must be done before stroking.

setLineColor :: (RealFloat a, Color c) => c -> Render a

Sets the stroke colour in current scope. If colour has opacity < 1, the scope opacity is set accordingly. Must be done before stroking.

setLineOpacity :: RealFloat n => n -> Render n

Sets the stroke opacity for the current scope. Should be a value between 0 and 1. Must be done before stroking.

Fill Options

setFillColor :: Color c => c -> Render n

Sets the fill colour for current scope. If an alpha colour is used, the fill opacity is set accordingly. Must be done before filling.

setFillRule :: FillRule -> Render n

Set the fill rule to winding or even-odd for current scope. Must be done before filling.

setFillOpacity :: RealFloat a => a -> Render n

Sets the stroke opacity for the current scope. Should be a value between 0 and 1. Must be done before stroking.

Transformations

setTransform :: RealFloat n => Transformation V2 n -> Render n

Resets the transform and sets it. Must be set before the path is used.

applyTransform :: RealFloat n => Transformation V2 n -> Render n

Applies a transformation to the current scope. This transformation only effects coordinates and text, not line withs or dash spacing. (See applyDeepTransform). Must be set before the path is used.

baseTransform :: RealFloat n => Transformation V2 n -> Render n

Base transforms are applied by the document reader.

applyScale :: RealFloat n => n -> Render n

resetNonTranslations :: Render n

Shading

linearGradient :: RealFloat n => Path V2 n -> LGradient n -> Render n

radialGradient :: RealFloat n => Path V2 n -> RGradient n -> Render n

colorSpec :: RealFloat n => n -> [GradientStop n] -> Render n

shadePath :: RealFloat n => Angle n -> Render n -> Render n

opacityGroup :: RealFloat a => a -> Render n -> Render n

images

image :: RealFloat n => DImage n External -> Render n

Images are wraped in a pgftext.

embeddedImage :: RealFloat n => DImage n Embedded -> Render n

embeddedImage' :: RealFloat n => ByteString -> Int -> Int -> T2 n -> Render n

Text

renderText :: [Render n] -> Render n -> Render n

setTextAlign :: RealFloat n => TextAlignment n -> [Render n]

Returns a list of values to be put in square brackets like pgftext[left,top]{txt}.

setTextRotation :: RealFloat n => Angle n -> [Render n]

setFontWeight :: FontWeight -> Render n

Set the font weight by rendering bf . Nothing is done for normal weight.

setFontSlant :: FontSlant -> Render n

Set the font slant by rendering bf . Nothing is done for normal weight.